+++ src/main/java/com/drathonix/loadmychunks/common/config/regis/RegistryRetriever.java
@@ -10,8 +10,26 @@ public abstract class RegistryRetriever<T> {
     private T object;
 
     public RegistryRetriever(String key){
-        this(ModResource.parse(key.replaceFirst("\\.",":")));
+        this(ModResource.parse(convertOldPatch(key)));
     }
+
+    /**
+     * In past versions of LMC a different file format was used and I had to keep colons out of strings. This fixes that.
+     * @param val original value
+     * @return possibly converted key.
+     */
+    private static String convertOldPatch(String val){
+        for (int i = 0; i < val.length(); i++) {
+            if(val.charAt(i) == ':'){
+                return val;
+            }
+            else if(val.charAt(i) == '.'){
+                return val.replaceFirst("\\.",":");
+            }
+        }
+        return val;
+    }
+
 
     public RegistryRetriever(ResourceLocation location) {
         this.location=location;
@@ -20,7 +38,7 @@ public abstract class RegistryRetriever<T> {
     abstract T retrieve();
 
     public String serializable() {
-        return location.toString().replaceFirst(":",".");
+        return location.toString();
     }
 
     public T get() {