+++ src/main/java/com/drathonix/loadmychunks/fabric/integ/CCTFabric.java
@@ -1,5 +1,5 @@
 //? if fabric && cc-tweaked {
-/*package com.drathonix.loadmychunks.fabric.integ;
+package com.drathonix.loadmychunks.fabric.integ;
 
 import com.drathonix.loadmychunks.common.LoadMyChunks;
 import com.drathonix.loadmychunks.common.integ.cct.CCTRegistryContent;
@@ -10,10 +10,10 @@ import com.drathonix.loadmychunks.common.registry.LMCContent;
 import dan200.computercraft.api.peripheral.PeripheralLookup;
 import dan200.computercraft.api.turtle.ITurtleUpgrade;
 //? if <1.20.4 {
-/^import dan200.computercraft.api.turtle.TurtleUpgradeSerialiser;
-^///?} else if <1.20.6 {
-/^import dan200.computercraft.api.upgrades.UpgradeSerialiser;
-^///?} else {
+/*import dan200.computercraft.api.turtle.TurtleUpgradeSerialiser;
+*///?} else if <1.20.6 {
+/*import dan200.computercraft.api.upgrades.UpgradeSerialiser;
+*///?} else {
 import dan200.computercraft.api.upgrades.UpgradeType;
 //?}
 import dev.architectury.registry.registries.DeferredRegister;
@@ -34,19 +34,19 @@ public class CCTFabric {
         PeripheralLookup.get().registerForBlockEntity((blockEntity, direction) -> new ChunkLoaderPeripheral(blockEntity.getBlockPos(),blockEntity.getLevel(), blockEntity.loadMyChunks$getChunkLoader()), LMCContent.chunkLoaderBlockEntity.get());
     }
     //?} else {
-    /^public static void init(){
+    /*public static void init(){
         //? if >1.20.1 {
         DeferredRegister<UpgradeSerialiser<? extends ITurtleUpgrade>> turtleUpgrades = DeferredRegister.create(LoadMyChunks.MOD_ID,ITurtleUpgrade.serialiserRegistryKey());
         //?} else {
-        /^¹DeferredRegister<TurtleUpgradeSerialiser<?>> turtleUpgrades = DeferredRegister.create(LoadMyChunks.MOD_ID,TurtleUpgradeSerialiser.registryId());
-        ¹^///?}
+        /^DeferredRegister<TurtleUpgradeSerialiser<?>> turtleUpgrades = DeferredRegister.create(LoadMyChunks.MOD_ID,TurtleUpgradeSerialiser.registryId());
+        ^///?}
         LMCContent.chunkLoaderBlockMap.forEach((color,supplier)->{
             TurtleChunkLoaderUpgrade tclu = new TurtleChunkLoaderUpgrade(supplier);
             //? if >=1.20.4 {
             RegistrySupplier<UpgradeSerialiser<? extends ITurtleUpgrade>> reg = turtleUpgrades.register((!color.isEmpty() ? color + "_" : "") + "chunk_loader", ()->UpgradeSerialiser.simple((key)->tclu));
             //?} else {
-            /^¹RegistrySupplier<TurtleUpgradeSerialiser<? extends ITurtleUpgrade>> reg = turtleUpgrades.register((!color.isEmpty() ? color + "_" : "") + "chunk_loader", ()-> TurtleUpgradeSerialiser.simple((key)->tclu));
-            ¹^///?}
+            /^RegistrySupplier<TurtleUpgradeSerialiser<? extends ITurtleUpgrade>> reg = turtleUpgrades.register((!color.isEmpty() ? color + "_" : "") + "chunk_loader", ()-> TurtleUpgradeSerialiser.simple((key)->tclu));
+            ^///?}
             CCTRegistryContent.registrySuppliers.add(reg);
         });
         turtleUpgrades.register();
@@ -55,10 +55,10 @@ public class CCTFabric {
         PeripheralLookup.get().registerForBlockEntity((blockEntity, direction) -> new ChunkLoaderPeripheral(blockEntity.getBlockPos(),blockEntity.getLevel(), blockEntity.loadMyChunks$getChunkLoader()), LMCContent.chunkLoaderBlockEntity.get());
 
     }
-    ^///?}
+    *///?}
 
     public static void clientInit(){
         CCTRegistryContent.registerClient();
     }
 }
-*///?}
+//?}